Extract GoogleTestMutationOperatorFilter.
