+++ include/GoogleTest/GoogleTestMutationOperatorFilter.h
@@ -1,0 +1,10 @@
+#include "MutationOperatorFilter.h"
+
+namespace mull {
+
+class GoogleTestMutationOperatorFilter : public MutationOperatorFilter {
+public:
+  bool shouldSkipInstruction(llvm::Instruction *instruction);
+};
+
+}
