+++ lib/CMakeLists.txt
@@ -26,6 +26,7 @@ set(mull_sources
   SimpleTest/SimpleTestRunner.cpp
 
   GoogleTest/GoogleTest_Test.cpp
+  GoogleTest/GoogleTestMutationOperatorFilter.cpp
   GoogleTest/GoogleTestFinder.cpp
   GoogleTest/GoogleTestRunner.cpp
