+++ lib/GoogleTest/GoogleTestMutationOperatorFilter.cpp
@@ -1,0 +1,23 @@
+#include "GoogleTestMutationOperatorFilter.h"
+
+#include "llvm/IR/DebugInfoMetadata.h"
+#include "llvm/IR/Instructions.h"
+
+using namespace llvm;
+using namespace mull;
+
+bool GoogleTestMutationOperatorFilter::shouldSkipInstruction(llvm::Instruction *instruction) {
+  if (instruction->hasMetadata()) {
+    int debugInfoKindID = 0;
+    MDNode *debug = instruction->getMetadata(debugInfoKindID);
+
+    DILocation *location = dyn_cast<DILocation>(debug);
+    if (location) {
+      if (location->getFilename().str().find("include/c++/v1") != std::string::npos) {
+        return true;
+      }
+    }
+  }
+
+  return false;
+};
